+++ APIBlox.AspNetCore.DynamicControllers/Types/DynamicControllerTemplateOptions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using APIBlox.AspNetCore.Enums;
 using APIBlox.NetCore.Extensions;
 
 namespace APIBlox.AspNetCore.Types